requirements.txt
# API clients
assemblyai>=0.37.0
anthropic>=0.39.0
PyGithub>=2.5.0
slack-sdk>=3.33.0

# ML / NLP
sentence-transformers>=3.3.0
scikit-learn>=1.5.0
numpy>=1.26.0
# faiss-cpu is not required: we use a NumPy cosine matrix.
# Uncomment the next line if you want to swap it back in.
# faiss-cpu>=1.9.0

# Graph
networkx>=3.4
pyvis>=0.3.2

# Backend
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.1
pydantic>=2.10.0
requests>=2.32.0
python-multipart>=0.0.20

# Frontend
streamlit>=1.40.0
plotly>=5.24.0
pandas>=2.2.0
